Реализовал паттерн state и закончил работу над шимером
