+++ RecipesAppTeamPlate/RecipesAppTeamPlate/RecipesModule/Model/RecipesStorage.swift
@@ -1,0 +1,12 @@
+// RecipesStorage.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// хранилище для ячейки Рыба
+struct RecipesStorage {
+    var dishLabel: String
+    var dishImage: String
+    var dishTime: String
+    var dishKkal: String
+}
