+++ RecipesAppTeamPlate/RecipesAppTeamPlate/RecipesModule/View/DetailRecipesModule/Model/StateShimer.swift
@@ -1,0 +1,14 @@
+// StateShimer.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Состояние шимера
+enum StateShimer {
+    /// инициализация
+    case initial
+    /// старт шиммера
+    case shimmer
+    /// данные загружены
+    case loaded
+}
